[NEW] Add RamDisk component
This component acts like a disk driver (secondary storage), but in reality
stores the data in the volatile memory area (RAM).

It shall be used, when relatively fast storage is needed and it is OK for the
data to be lost between reset cycles. Use cases are for example caching or
testing.

Fixes:
SEOS-1092