[REF] Change type of the offset to off_t
When accessing storage we need to be able to address even large storages, so
the offset must be of an off_t type.

Note that offset is signed, so driver must report an error if an operation with
the negative offset was requested.

In case of `erase` size argument has been changed as well, as we want to be
able to erase large areas at one shot.

Fixes:
SEOS-1472