[BLD] Build SDK binaries as a RelWithDebInfo
Since we want to only allow building certain build types, we need to
select one for our SDK binaries (proxy app, etc.).

Release with debugging info is a good choice as in a result we will
have binaries that are at the same time optimized and debug-able, which
might be useful when analyzing test framework related issues.

See:
SEOS-2624