[DEV] Parallelize unit-test, astyle and sanity-check
Because the aforementioned steps are independent of each other they are now executed in parallel. Unit tests and astyle are using the development SDK package whereas the sanity check is using the SDK package.

To make this work some astyle scripts are now removed only for the final SDK package and the unit tests work on a pre-created package and do not collect the sources by themselves anymore.