[DEV] Build SDK demos against sdk-package instead of repo
If a test system is included in the DEMO_LIST it is now built against the SDK package and not checked out from the default repo (e.g. "ss/demo_hello_world"). For test systems not part of DEMO_LIST the non-default repo functionality ("name@repo/path") is still supported.