[DEV] Use flag to identify SDK demos
Instead of checking whether TEST_SYSTEM starts with "demos/" now a new flag is used. This requires a change in the Jenkins job configuration in order to hand over the variable properly.

Using this flag also the displayed text has been improved to also print "demo" for SDK demos (in addition to "build" and "test").